+++ crates/runc-shim/Cargo.toml
@@ -31,6 +31,7 @@ serde = { version = "1.0.133", features = ["derive"] }
 serde_json = "1.0.74"
 oci-spec = "0.6.0"
 crossbeam = "0.8.1"
+uuid = { version = "1.0.0", features = ["v4"] }
 
 # Async dependencies
 async-trait = { workspace = true, optional = true }
