+++ crates/runc-shim/src/synchronous/console.rs
@@ -14,16 +14,16 @@
    limitations under the License.
 */
 
-use std::path::{Path, PathBuf};
+use std::{
+    os::unix::net::{UnixListener, UnixStream},
+    path::{Path, PathBuf},
+};
 
+use containerd_shim::{io_error, util::mkdir, Error, Result};
 use log::warn;
-use tokio::net::{UnixListener, UnixStream};
 use uuid::Uuid;
 
-use crate::{
-    util::{mkdir, xdg_runtime_dir},
-    Error, Result,
-};
+use crate::common::xdg_runtime_dir;
 
 pub struct ConsoleSocket {
     pub listener: UnixListener,
@@ -32,9 +32,9 @@ pub struct ConsoleSocket {
 }
 
 impl ConsoleSocket {
-    pub async fn new() -> Result<ConsoleSocket> {
+    pub fn new() -> Result<ConsoleSocket> {
         let dir = format!("{}/pty{}", xdg_runtime_dir(), Uuid::new_v4());
-        mkdir(&dir, 0o711).await?;
+        mkdir(&dir, 0o711)?;
         let file_name = Path::new(&dir).join("pty.sock");
         let listener = UnixListener::bind(&file_name).map_err(io_error!(
             e,
@@ -48,29 +48,23 @@ impl ConsoleSocket {
         })
     }
 
-    pub async fn accept(&self) -> Result<UnixStream> {
-        let (stream, _addr) = self
-            .listener
-            .accept()
-            .await
-            .map_err(io_error!(e, "failed to list console socket"))?;
+    pub fn accept(&self) -> std::io::Result<UnixStream> {
+        let (stream, _addr) = self.listener.accept()?;
         Ok(stream)
     }
+}
 
-    // async drop is not supported yet, we can only call clean manually after socket received
-    pub async fn clean(self) {
+impl Drop for ConsoleSocket {
+    fn drop(&mut self) {
         if self.rmdir {
-            if let Some(tmp_socket_dir) = self.path.parent() {
-                tokio::fs::remove_dir_all(tmp_socket_dir)
-                    .await
-                    .unwrap_or_else(|e| {
-                        warn!(
-                            "remove tmp console socket path {} : {}",
-                            tmp_socket_dir.display(),
-                            e
-                        )
-                    })
-            }
+            let tmp_socket_dir = self.path.parent().unwrap();
+            std::fs::remove_dir_all(tmp_socket_dir).unwrap_or_else(|e| {
+                warn!(
+                    "remove tmp console socket path {} : {}",
+                    tmp_socket_dir.to_str().unwrap(),
+                    e
+                )
+            })
         }
     }
 }
