+++ crates/shim/Cargo.toml
@@ -29,7 +29,6 @@ time = { version = "0.3.7", features = ["serde", "std"] }
 serde_json = "1.0.78"
 serde_derive = "1.0.136"
 serde = "1.0.136"
-uuid = { version = "1.0.0", features = ["v4"] }
 signal-hook = "0.3.13"
 oci-spec = "0.6.0"
 prctl = "1.0.0"
