+++ crates/shim/src/util.rs
@@ -15,7 +15,6 @@
 */
 
 use std::{
-    env,
     os::unix::io::RawFd,
     time::{SystemTime, UNIX_EPOCH},
 };
@@ -164,13 +163,6 @@ pub fn convert_to_any(obj: Box<dyn MessageDyn>) -> Result<Any> {
     any.type_url = obj.descriptor_dyn().full_name().to_string();
 
     Ok(any)
-}
-
-/// Returns a temp dir. If the environment variable "XDG_RUNTIME_DIR" is set, return its value.
-/// Otherwise if `std::env::temp_dir()` failed, return current dir or return the temp dir depended on OS.
-pub(crate) fn xdg_runtime_dir() -> String {
-    env::var("XDG_RUNTIME_DIR")
-        .unwrap_or_else(|_| env::temp_dir().to_str().unwrap_or(".").to_string())
 }
 
 pub trait IntoOption
